refactor(perf): mv pkg/performance/ringbuffer -> pkg/ringbuffer
