+++ pkg/performance/collectors/kernel.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/antimetal/agent/pkg/performance"
 	"github.com/antimetal/agent/pkg/performance/capabilities"
-	"github.com/antimetal/agent/pkg/performance/ringbuffer"
+	"github.com/antimetal/agent/pkg/ringbuffer"
 	"github.com/antimetal/agent/pkg/proc"
 	"github.com/go-logr/logr"
 )
